fix: resolve int types to uint types

+++ consensus/consensus.go
@@ -12,16 +12,16 @@ type ConsensusAlgorithm interface {
 	SelectValidator() common.Address
 
 	// GetReward calculates the reward for the validator/miner.
-	GetReward() int
+	GetReward() uint64
 
 	// Deposit deposits a validator's stake.
-	Deposit(validator common.Address, amount int)
+	Deposit(validator common.Address, amount uint64)
 
 	// Withdraw withdraws a validator's stake.
-	Withdraw(validator common.Address, amount int)
+	Withdraw(validator common.Address, amount uint64)
 
 	// GetValidatorStake returns the stake of a validator.
-	GetValidatorStake(validator common.Address) int
+	GetValidatorStake(validator common.Address) uint64
 
 	// GetValidatorSet returns the current validator set.
 	GetValidatorSet() []common.Address
@@ -31,7 +31,7 @@ type ConsensusAlgorithm interface {
 
 	// CalculateValidatorReward calculates the reward for a specific validator
 	// based on their performance metrics and status.
-	CalculateValidatorReward(validator common.Address) int
+	CalculateValidatorReward(validator common.Address) uint64
 
 	// RecordBlockProduction records that a validator successfully produced a block
 	RecordBlockProduction(validator common.Address)
